🔗 Connect half prior drafting
Still deciding on whether or not I think direct hyperlinks should be used much - what do you think? 🧐

Also, on [^american] and other rhetoric that surrounds my work: I understand that these are fundamentally global challenges (that all fall on the same challenges), but we are writing to US regulators here - so I leverage US arguments. 🇺🇸